# ARROW-12709: [C++] Add binary_join_element_wise