… cpp/src/arrow/compute/kernels/scalar_compare_test.cc
… cpp/src/arrow/compute/kernels/scalar_string_test.cc
… docs/source/cpp/compute.rst
… python/pyarrow/_compute.pyx
… python/pyarrow/includes/libarrow.pxd
… python/pyarrow/tests/test_compute.py