@ianmcook 2021-06-14
Thanks for working on this @lidavidm!

I will add the relevant functions to the R bindings after this is merged (ARROW-11514).