@jorisvandenbossche 2021-06-14
Some naming nitpicks ;) 
I think "var_args_join" is not super clear. Having a notion about it being for string data would be good, and the scalar list of string join kernel that was just added in ARROW-10959 is called "binary_join" (binary because it supports all binary and not just string, which is the same here I think). So something about "binary_join_var_args" ? 
Another reference is the variadic element-wise min/max kernels that were added, where "element_wise" was used and not "var_args". That would then give something like "binary_join_element_wise" (a mouthful ..)