@lidavidm 2021-06-14
Or perhaps `element_wise_binary_join` since it's also `element_wise_min`?