@jorisvandenbossche 2021-06-14
That would indeed be more consistent. 
Personally, searching for the function / "tab-completion" in mind, I think having the name start with "binary_join" or "string_join" is more useful (I could say the same for the min/max version though). 
We could also rename the just merged "binary_join" to "binary_join_list" to more clearly differentiate the different kernels.