@lidavidm 2021-06-14
Ah, maybe then we should rename `element_wise_min` to `min_element_wise`.