@ianmcook 2021-06-14
I'm unclear on what the intended behavior is when you choose the {{SKIP}} null handling behavior and the separator is an array. Could you describe that please? Thanks