@lidavidm 2021-06-14
The null handling behavior never affects the separator: it only describes how to handle the other values. It's intended to let you mimic libcudf.

```
binary_join_element_wise(['a', 'b', 'c', 'h'], ['d', null, 'e', 'i'], ['f', 'g', null, 'j'], [';', '-', ',', null])
EMIT_NULL:                ['a', null, null, null]
SKIP:                     ['a;d;f', 'b-g', 'c,e', null] # b and g are joined; the middle null value is ignored
REPLACE(replacement=foo): ['a;d;f', 'b-foo-g', 'c,e,foo', null]
```