@ianmcook 2021-06-14
Oh, I see; I was confused about what happens when the separator is an array; I see now. Thank you!