@jorisvandenbossche 2021-06-15
> The use of "binary" in the names of these string join kernels is unfortunate; it's not clear at first glance whether "binary" is a reference to the arity or to the input type.

Yeah, personally I would be fine to have an alias that uses "string" to have some more recognizable name ("string_join_element_wise/list"), but not sure we want to start adding aliases in general ..