@pitrou 2021-06-15
I'm not really surprised and was actually expecting worse results :-) There are two factors:
* the reshuffling (row-oriented vs. column-oriented) is an intrinsic problem of this operation
* the implementation iterates row a time, and therefore does dynamic switching between various cases (scalar, array, null options...) in each loop iteration