@pitrou 2021-06-15
An implementation iterating over each column in turn might be more performant, but at the cost of more complexity:
* you first need to compute the output null bitmap and offsets
* you need to keep a vector of output pointers into the output data (a bit similar to the CSV writer)