@lidavidm 2021-06-15
I was just thinking that. The current implementation already computes the output lengths (accounting for nullability) up front, so, it might not be too bad.

I also noticed the current implementation effectively branches on scalar/array twice per column in the loop, trying to optimize that improved the cases with fewer columns at the expense of those with more columns (though that may be the right tradeoff). But I'll give the columnwise approach a try.